requirements.txt
click
wikipedia
pytube

pydub
openai
tiktoken

httpx
torch
numpy
protobuf<=3.19.0
sentence-transformers

PyPDF2
qdrant_client
pydantic
python-telegram-bot
tenacity